% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.sim.conditionalPoisson.R
\name{mf.sim.condPoisson}
\alias{mf.sim.condPoisson}
\title{Simulate data from Poisson distribution and calculate coefficients}
\usage{
mf.sim.condPoisson(vect.margprob_GS_1, vect.margprob_GS_0, mat.cov_GS_1,
  mat.cov_GS_0, n.sim, sim.N_1, sim.N_0)
}
\arguments{
\item{vect.margprob_GS_1}{<object; input>}

\item{vect.margprob_GS_0}{<object; input>}

\item{mat.cov_GS_1}{<object; input>}

\item{mat.cov_GS_0}{<object; input>}

\item{n.sim}{<numeric; proccessing>}

\item{sim.N_1}{<numeric; proccessing>}

\item{sim.N_0}{<numeric; proccessing>}
}
\description{
Simulate data from Poisson distribution and calculate coefficients
}
